# ipx/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Solana transaction API",
        "title": "SVMlab API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:33153",
    "basePath": "/",
    "paths": {
        "/svm/account/airdrop": {
            "post": {
                "description": "Requests a fixed half a SOL. The amount is not a field because the faucet enforces its own limit, and asking above it fails the call rather than handing out less.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Fund an account on devnet or testnet",
                "parameters": [
                    {
                        "description": "Account",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.AirdropRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/account.AirdropResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/account/balance": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Read an account's balance",
                "parameters": [
                    {
                        "description": "Account",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.AccountRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/account.BalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/account/nonce": {
            "post": {
                "description": "Parses the 80 bytes a nonce account holds. The stored nonce is what a transaction may carry in place of a recent blockhash so that it never expires, which is the opposite of an EVM nonce: there a per-account counter prevents replay, here expiry is what is being removed. An account sized by create-account but never initialized is a normal intermediate state and reports initialized=false with all-zero fields.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Read a durable nonce account",
                "parameters": [
                    {
                        "description": "Nonce account",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.NonceRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/account.NonceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/account/owner": {
            "post": {
                "description": "Only the owning program may debit an account or write its data. An account owned by anything other than the System Program cannot be moved with a system transfer, and one owned by a non-executable address cannot be moved at all, so system_owned answers whether the balance is still reachable.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Read who owns an account",
                "parameters": [
                    {
                        "description": "Account",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.AccountOwnerRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/account.AccountOwnerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/account/state": {
            "post": {
                "description": "Returns the owner, executable flag, data size, balance, and data. An account that does not exist is reported with exists=false rather than as an error, since most 32-byte values name one nobody has created.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Read an account's on-chain state",
                "parameters": [
                    {
                        "description": "Account",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.AccountRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/account.AccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/account/tokens": {
            "post": {
                "description": "Lists token accounts owned by a wallet under both the classic Token Program and Token-2022. The response keeps the program on each account because the two programs own separate accounts and no single token instruction can touch both at once.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "List the token accounts a wallet owns",
                "parameters": [
                    {
                        "description": "Wallet",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.TokensRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/account.TokensResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "account.AccountOwnerRequest": {
            "type": "object",
            "properties": {
                "public_key": {
                    "type": "string",
                    "example": "EodYvwsT22JTdNmvCeC974WjPiVYcxvfGpYLJxnB3JqK"
                }
            }
        },
        "account.AccountOwnerResponse": {
            "type": "object",
            "properties": {
                "exists": {
                    "type": "boolean"
                },
                "owner": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "system_owned": {
                    "type": "boolean"
                }
            }
        },
        "account.AccountRequest": {
            "type": "object",
            "properties": {
                "public_key": {
                    "type": "string",
                    "example": "EodYvwsT22JTdNmvCeC974WjPiVYcxvfGpYLJxnB3JqK"
                }
            }
        },
        "account.AccountResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "executable": {
                    "type": "boolean"
                },
                "exists": {
                    "type": "boolean"
                },
                "lamports": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                },
                "space": {
                    "type": "integer"
                }
            }
        },
        "account.AirdropRequest": {
            "type": "object",
            "properties": {
                "public_key": {
                    "type": "string"
                }
            }
        },
        "account.AirdropResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                }
            }
        },
        "account.BalanceResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                }
            }
        },
        "account.NonceRequest": {
            "type": "object",
            "properties": {
                "public_key": {
                    "type": "string",
                    "example": "EodYvwsT22JTdNmvCeC974WjPiVYcxvfGpYLJxnB3JqK"
                }
            }
        },
        "account.NonceResponse": {
            "type": "object",
            "properties": {
                "authority": {
                    "type": "string"
                },
                "exists": {
                    "type": "boolean"
                },
                "initialized": {
                    "type": "boolean"
                },
                "lamports_per_signature": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "account.TokenAccountResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "close_authority": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "delegate": {
                    "type": "string"
                },
                "delegated_amount": {
                    "type": "string"
                },
                "exists": {
                    "type": "boolean"
                },
                "mint": {
                    "type": "string"
                },
                "native": {
                    "type": "boolean"
                },
                "owner": {
                    "type": "string"
                },
                "program": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "rent_reserve": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "account.TokensRequest": {
            "type": "object",
            "properties": {
                "public_key": {
                    "type": "string",
                    "example": "EodYvwsT22JTdNmvCeC974WjPiVYcxvfGpYLJxnB3JqK"
                }
            }
        },
        "account.TokensResponse": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/account.TokenAccountResponse"
                    }
                },
                "count": {
                    "type": "integer"
                },
                "public_key": {
                    "type": "string"
                }
            }
        },
...
    }
}